fix errors after rebasing to origin/master
